requirements.txt
flask>=2.0.1
mysql-connector-python>=8.0.26
requests>=2.26.0
schedule>=1.1.0
psutil>=5.9.0